requirements.txt
requests
pandas
yfinance
python-dotenv
streamlit
plotly
google-generativeai
schedule
streamlit-authenticator
cryptography
bcrypt
